Initial effort to switch to the purecss framework. We have a pretty rubbish looking web page for now but all traces of Bulma have been removed.
